+++ juniper/src/schema/mod.rs
@@ -1,3 +1,4 @@
 pub mod meta;
 pub mod model;
 pub mod schema;
+pub mod translate;